refactor: MemberNoGenerator util 패키지로 이동 (#123)
